hello world

+++ libfs/fs.c
@@ -1409,7 +1409,7 @@ int fs_read(int fd, void *buf, size_t count)
         free(bounce_buffer);
         return -1;
     }
-    printf("offset in fs_read = %d\n", offset);
+    printf("offset in fs_read = %d\n", (int)offset);
     memcpy(buf + buf_idx, bounce_buffer + offset % BLOCK_SIZE, clamp(real_count_temp, BLOCK_SIZE));
 
     buf_idx += clamp(real_count_temp, BLOCK_SIZE);
